Create SciWhy.py
This is part of a module which links science fiction novels which are mentioned in chat. Due to OS compatibility, it may need some tweaks before being incorporated into the main mod. The large JSON file it uses as a database makes it very slow to load, but the resulting performance is quite fast.